#132: Consolidating files, termgroup => termgroups, termset => termsets
