+++ src/sharepoint/taxonomy/termsets.ts
@@ -1,11 +1,11 @@
 "use strict";
 
 import { QueryableTaxonomy } from "./QueryableTaxonomy";
-import { TermSet } from "./TermSet";
 import { Util } from "../../utils/util";
 
 /**
- * Root of the SharePoint Taxonomy module
+ * Describes a collection of TermSet objects
+ *
  */
 export class TermSets extends QueryableTaxonomy {
     private groupIdentifier: string;
@@ -44,6 +44,48 @@ export class TermSets extends QueryableTaxonomy {
                     groups.getById(new SP.Guid(this.groupIdentifier)) :
                     groups.getByName(this.groupIdentifier);
                 this.clientObjects = group.get_termSets();
+                this.taxSession.retrieveObjects(this.clientContext, this).then(resolve, reject);
+            });
+        });
+    }
+};
+
+/**
+ * Describes a single TermSet instance
+ *
+ */
+export class TermSet extends QueryableTaxonomy {
+    private groupIdentifier: string;
+    private identifier: string;
+
+    /**
+     * Creates a new instance of the TermSet class
+     */
+    constructor(groupIdentifier: string, identifier: string) {
+        super("SP.Taxonomy.TermSet");
+        this.groupIdentifier = groupIdentifier;
+        this.identifier = identifier;
+    }
+
+    public terms(): Terms {
+        return new Terms(this.groupIdentifier, this.identifier, "get_terms");
+    }
+
+    public allterms(): Terms {
+        return new Terms(this.groupIdentifier, this.identifier, "getAllTerms");
+    }
+
+    public get(): Promise<any> {
+        return new Promise((resolve, reject) => {
+            this.taxSession.getDefaultSiteCollectionTermStore(this.clientContext).then(defaultTermStore => {
+                let groups = defaultTermStore.get_groups();
+                let group = Util.isValidGUID(this.groupIdentifier) ?
+                    groups.getById(new SP.Guid(this.groupIdentifier)) :
+                    groups.getByName(this.groupIdentifier);
+                let termSets = group.get_termSets();
+                this.clientObjects = Util.isValidGUID(this.identifier) ?
+                    termSets.getById(new SP.Guid(this.identifier)) :
+                    termSets.getByName(this.identifier);
                 this.taxSession.retrieveObjects(this.clientContext, this).then(resolve, reject);
             });
         });
